serviceproxy/gateway_on_streamlit/requirements.txt
google-generativeai>=0.8.3,<1
openai>=1.54.5,<2
pydantic>=2.7.0,<3
python-dotenv>=1.0.1,<2
streamlit>=1.40.2
